+++ azext_iot/sdk/digitaltwins/controlplane/azure_digital_twins_management_client.py
@@ -88,7 +88,7 @@ class AzureDigitalTwinsManagementClient(SDKClient):
         super(AzureDigitalTwinsManagementClient, self).__init__(self.config.credentials, self.config)
 
         client_models = {k: v for k, v in models.__dict__.items() if isinstance(v, type)}
-        self.api_version = '2022-10-31'
+        self.api_version = '2023-01-31'
         self._serialize = Serializer(client_models)
         self._deserialize = Deserializer(client_models)